requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0

# Evaluation metrics
nltk>=3.8
bert-score>=0.3.13
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Configuration
hydra-core>=1.3.0
pyyaml>=6.0

# Data processing
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# Optional: Experiment tracking
# wandb>=0.15.0

# Optional: Fast inference
# vllm>=0.2.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0